fixed a pretty big bug with the magnetic field:
basically, we were updating the points of the line based on the center value,
but were accidentally recalculating this center point after changing the left-point's
position, which was changing the center so that the second point was, in turn, wrong.

ie we were doing this:

x = f(x,y);
y = f(x,y);

when we should have done this:

xtmp = x;
ytmp = y;
x = f(xtmp, ytmp);
y = f(xtmp, ytmp);

yeah. :(